fix: 🐛 allow changing subscription type when canceling checkout
Previously, the cached checkout cart was always used for checkouts,
which means the user's changed subscription type doesn't get reflected.
This commit fixes two things: 1. a different checkout cart is created if
the cached one doesn't match the type. 2. the user's subscription type
is remembered in the purchase portal screen as a coutesy.

We also handle refreshing of an expired/nearly-expired checkout session.